converts Spice_style.odt to asciidoc spice_style.txt
The conversion try keeps the content as is no matter if updated or not.
Asciidoc format is already used for manual.
Using a text format make easier to see git diff and send patches.
Also it's easier to convert to different format and copy to web pages.

Related: https://bugs.freedesktop.org/show_bug.cgi?id=95258

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>
Acked-by: Pavel Grunt <pgrunt@redhat.com>