RedsState: clean up spice_server_char_device_add_interface
Previously we were creating a variable named 'dev_state' and then
apparently not using it. Well, we *were* actually using it, but in a
convoluted sort of way. Creating a new RedCharDevice has a
side-effect of setting itself as the 'st' attribute of
SpiceCharDeviceInstance. So 'dev_state' and 'char_device->st' are in
fact the same variable. But they were being used interchangeably, which
was rather confusing. For example

if (dev_state)
    // do something with char_device->st

So this patch doesn't actually change anything, but it makes the code a
bit easier to follow.

Acked-by: Frediano Ziglio <fziglio@redhat.com>